+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- `value` field attribute form `Enum` derive macro;
+
 ## [5.1.0] - 2025-10-29
 
 ### Changed